Rename currency-query-params KT-4
