+++ src/core/models/index.ts
@@ -1,3 +1,3 @@
-export * from './currency-query-params';
+export * from './query-params';
 export * from './exchange-rates';
 export * from './currency';
