+++ src/core/models/query-params.ts
